requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
transformers>=4.35.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
faiss-cpu>=1.7.4
neo4j>=5.14.0
py2neo>=2021.2.3
networkx>=3.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tqdm>=4.66.0
python-dotenv>=1.0.0
datasets>=2.14.0
accelerate>=0.24.0
jupyter>=1.0.0
ipykernel>=6.26.0